modificacao dos diretorios
