--- Senac Libras (register)/static/register.css
@@ -1,3 +1,0 @@
-.container{
-    background-color: #F6F6F6;
-}
