+++ static/register.css
@@ -1,0 +1,3 @@
+.container{
+    background-color: #f9f9f9;
+}
